1. First version that runs ./build.sh test all using Eigen vectors for testing. 2. Set BOOST_INCLUDE in configure.ac instead of each makefile.am. 3. Define and use EIGEN_INCLUDE like BOOST_INCLUDE. 4. Fix more uses of SimpleVector.size() that need conversion to size_t. 5. Change autoconfig conditional CppAD_EIGEN to CppAD_EIGEN_DIR.
git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2453 0e562018-b8fb-0310-a933-ecf073c7c197

+++ cppad/local/sparse_jacobian.hpp
@@ -927,7 +927,7 @@ void ADFun<Base>::SparseJacobianCase(
 	CheckSimpleVector<Base, VectorBase>();
 
 	CPPAD_ASSERT_KNOWN(
-		p.size() == m * n,
+		size_t(p.size()) == m * n,
 		"SparseJacobian: using bool values and size of p "
 		" not equal range dimension times domain dimension for f"
 	);
